backend/requirements.txt
alembic
aniso8601==8.0.0
appdirs==1.4.3
attrs==19.3.0
Babel==2.7.0
bcrypt==3.1.7
beautifulsoup4==4.8.1
black==19.3b0
blinker==1.4
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
Django==2.2.4
django-appconf==1.0.3
dominate==2.3.5
Flask==1.0.2
Flask-Admin==1.5.3
Flask-Babel==0.12.2
Flask-BabelEx==0.9.3
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Generic-Views==0.1.1
Flask-HTTPAuth==3.3.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.8.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.0
Flask-Testing==0.7.1
Flask-WTF==0.14.2
gunicorn==19.9.0
httpie==1.0.2
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.12
MarkupSafe==1.1.1
marshmallow==2.20.1
mysql==0.0.2
mysql-connector-python==8.0.16
mysqlclient==1.4.4
numpy==1.17.3
pandas==0.25.2
passlib==1.7.1
peewee==3.9.6
pinax-testimonials==2.0.4
protobuf==3.8.0
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.4
speaklater==1.3
SQLAlchemy==1.3.5
sqlparse==0.3.0
toml==0.10.0
urllib3==1.25.3
visitor==0.1.3
Werkzeug
WTForms==2.2.1
redis==2.10.6
rq==0.12.0
boto3==1.9.17
